+++ packages/brick_offline_first_with_rest/lib/src/offline_queue/rest_request_sqlite_cache_manager.dart
@@ -8,7 +8,7 @@ import 'package:sqflite_common/sqlite_api.dart' show DatabaseFactory;
 class RestRequestSqliteCacheManager extends RequestSqliteCacheManager<http.Request> {
   RestRequestSqliteCacheManager(
     String databaseName, {
-    DatabaseFactory? databaseFactory,
+    required DatabaseFactory databaseFactory,
     Duration? processingInterval,
     bool? serialProcessing,
   }) : super(