+++ packages/brick_offline_first_with_rest/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 
 * Import from `sqflite_common` instead of `sqflite` to avoid a Flutter dependency
+* **BREAKING CHANGE** `OfflineFirstWithRestRepository#offlineQueueHttpClientRequestSqliteCacheManager` is now `offlineQueueManager`
+* **BREAKING_CHANGE** `offlineQueueManager` is required to create `BrickOfflineFirstWithRestRepository`. To migrate without recreating the queue database, pass RestRequestSqliteCacheManager('brick_offline_queue.sqlite', databaseFactory)
 
 ## 1.0.0